+++ VoiceController/BulkRequest.cs
@@ -16,12 +16,12 @@ namespace VoiceController
         private string ringUrl = "";
         private string answerUrl = "";
         private string hangupUrl = "";
-        private short retries = (short)0;
+        private byte retries = 0;
         private string callerId = "";
-        private short status = (short)0;
+        private byte status = 0;
         private int processedCount = 0;
         private long voiceRequestId = 0L;
-        private short toolId = 0;
+        private byte toolId = 0;
 
         public long Id { get { return this.id; } set { this.id = value; } } 
         public string Xml { get { return this.xml; } set { this.xml = value; } } 
@@ -31,12 +31,12 @@ namespace VoiceController
         public string RingUrl { get { return this.ringUrl; } set { this.ringUrl = value; } } 
         public string AnswerUrl { get { return this.answerUrl; } set { this.answerUrl = value; } } 
         public string HangupUrl { get { return this.hangupUrl; } set { this.hangupUrl = value; } } 
-        public short Retries { get { return this.retries; } set { this.retries = value; } } 
+        public byte Retries { get { return this.retries; } set { this.retries = value; } } 
         public string CallerId { get { return this.callerId; } set { this.callerId = value; } } 
-        public short Status { get { return this.status; } set { this.status = value; } }
+        public byte Status { get { return this.status; } set { this.status = value; } }
         public int ProcessedCount { get { return this.processedCount; } set { this.processedCount = value; } } 
         public long VoiceRequestId { get { return this.voiceRequestId; } set { this.voiceRequestId = value; } }
-        public short ToolId { get { return this.toolId; } set { this.toolId = value; } }
+        public byte ToolId { get { return this.toolId; } set { this.toolId = value; } }
         public string DisplayString()
         {
             return " Id : " + this.Id.ToString() + ", RingUrl : " + this.RingUrl + ", AnswerUrl : " + this.AnswerUrl + ", HangupUrl : " + this.HangupUrl + ", CallerId : " + this.CallerId + ", Status : " + this.Status.ToString() + ", ProcessedCount : " + this.ProcessedCount.ToString() + ", VoiceRequestId : " + this.VoiceRequestId.ToString();
